test-scenarios/2.1-new-failing-testfile.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Headers structure</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#headers">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>

      test(function() {
        assert_equals(true, false, "this assertion should always fail");
      }, "Intentionally failing test scenario");

   </script>
  </body>
</html>
